fixing some -Wsign-compare stuff

+++ libraries/npy/npy.c
@@ -9,7 +9,7 @@
 
 static bool
 consume(FILE *f, char *str) {
-    for (int i = 0; i < strlen(str); i++) {
+    for (size_t i = 0; i < strlen(str); i++) {
         char c = fgetc(f);
         if (feof(f) || c != str[i]) {
             return false;
